requirements.txt
alembic==1.10.2
attrs==22.2.0
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
exceptiongroup==1.1.1
Fabric==1.15.0
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
invoke==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
paho-mqtt==1.6.1
paramiko==2.12.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
PyNaCl==1.5.0
pypdf==3.7.0
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.8
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
